+++ src/com/cab404/syncronos/TimelineObjectStorage.java
@@ -1,15 +1,22 @@
 package com.cab404.syncronos;
 
-import java.util.Collection;
 
 /**
- * Sorry for no comments!
+ * Specifies logic of data mining.
+ * Should be possibly done with trees.
  * Created at 4:39 on 23.02.15
  *
  * @author cab404
  */
 public interface TimelineObjectStorage {
 
-    public Iterable<TimelineObject> getObjects(long start, long end, float zoom);
+    /**
+     * Returns events in given time span
+     *
+     * @param start Start of time span
+     * @param end   End of time span
+     * @param zoom  Zoom of the map
+     */
+    public Iterable<TimelineObject> getObjects(long start, long end, double zoom);
 
 }
